Snake Game UPDATE 5 (MASSIVE UPDATE!)
Finally made the border collision. Wasn't that hard just reworked the food collision. Now when you eat the fruit the body part won't stay at the middle, it will follow the head. Tried to do something with the score text, but it not works, maybe in update 6 I will make it work like it needs to be.